dir: Allow downgrading when using collection IDs
Currently when the P2P code paths are used, the OstreeRepoFinderResult
objects returned by find_remotes_async() specify the latest commit
available for each ref from each remote, and that commit is then pulled.
This breaks downgrades because in that case a specific commit hash is
specified by the user but never pulled. Fix that by editing the
OstreeRepoFinderResult objects to point to the desired commit rather
than the latest one. This will likely fail for some remotes (because
it's a different commit than they advertised) but that is the desired
behavior.

Fixes https://github.com/flatpak/flatpak/issues/1309