dir: Allow downgrading when using collection IDs
Currently when the P2P code paths are used, there's no way to downgrade
an app to a specific commit because the libostree API doesn't support
it. This commit adds the requested commit checksums to the
`override-commit-ids` key in the options parameter of calls to
ostree_repo_find_remotes_async() so that once libostree supports that
the returned OstreeRepoFinderResults will point to that checksum rather
than the latest available one.

The effect is that `flatpak update --commit=HASH APP` will work when APP
comes from a repository that has a collection ID configured.

Fixes https://github.com/flatpak/flatpak/issues/1309