lib: Add P2P support to list_installed_refs_for_update()
Currently flatpak_insallation_list_installed_refs_for_update() only
checks the configured remotes for updates. This means that updates from
peer sources such as USB drives or computers on the LAN are never seen,
which is especially problematic when the computer is offline. The end
result is that GNOME Software doesn't show an "Update" button to let you
update an app from a USB drive when you're offline.

This commit uses ostree API, primarily ostree_repo_find_remotes_async(),
to check peer sources for updates. Since we use the recently added
"<remote>:" syntax for listing locally available remote refs, you must
use a version of ostree at or later than 2018.3 for this to work (for
the releases 2017.13 - 2018.2 ostree didn't support that or the
previously supported "<remote>:." syntax). The ostree requirement wasn't
bumped in configure.ac because ostree 2018.3 isn't yet in Fedora (but
this will only be used by Endless for now anyway). We're also using an
ostree 2018.5 feature from
https://github.com/ostreedev/ostree/pull/1518, but I used a version
check for that. Without it you can't tell if an offered commit is an
update or a downgrade, but that behavior is still better than the status
quo.