dir: Support filtering app installs by user’s OARS settings
Use the user’s OARS filter to prevent installation of apps which have
more extreme content than the user is allowed to see.

This uses libeos-parental-controls to load the user’s enforced OARS
filter, which describes the extremeness of each type of content the user
is allowed to see. If an app they are trying to install exceeds the
filter value in any OARS section, installation is disallowed and an
error is returned.

This adds a dependency on libappstream-glib to common/.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T24024

Rebase 1.1.3 (T24078): Revert the upstream commit to move
  flatpak_dir_load_appstream_store() to app/