Fetch ostree-metadata for backwards compat
On versions of Flatpak prior to 1.7.1, the offline app install functionality
depends on the "ostree-metadata" ref being available for each
collection. So for example a flash drive containing refs from both
eos-apps and flathub would have to have both
(com.endlessm.Apps, ostree-metadata)
and
(org.flathub.Stable, ostree-metadata)
on it. However Flatpak from 1.7.1 onward no longer depends on
ostree-metadata to do offline updates and therefore no longer pulls it
during online install/update operations (see
https://github.com/flatpak/flatpak/pull/3476).

Flatpak does still put ostree-metadata on a drive in "flatpak
create-usb" if it exists locally. So this is a patch to fill the gap:
pull ostree-metadata any time we're fetching from a remote, so it will
be available when create-usb is run, so such USB drives will work with
computers running older flatpak versions, such as Endless OS versions
older than 3.9.0 (which is planned for November 2020).

Also, update the unit tests to check that this is working.

This has the additional benefit of checking that the locally configured
collection ID matches the one configured on the server side. A mismatch
would cause offline distribution to fail.